requirements.txt
streamlit
opencv-python-headless
azure-ai-inference
requests
pandas
numpy
python-dotenv